feat: add sse transport example and server creator
