+++ package.json
@@ -54,7 +54,9 @@
   "dependencies": {
     "@inkjs/ui": "^2.0.0",
     "@modelcontextprotocol/sdk": "^1.10.2",
+    "@types/express": "^5.0.2",
     "@types/yargs": "^17.0.33",
+    "express": "^5.1.0",
     "ink": "^5.2.0",
     "node-notifier": "^10.0.1",
     "pino": "^9.6.0",
